pyproject.toml
[project]
name = "projeto-lacrei"
version = "0.1.0"
description = ""
authors = [
    {name = "Thais Calvi",email = "thaispinheiro@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
